Refactor private AI key models and enhance region tests
